[Feat/#3] Refactor: `RadialChart` 컴포넌트 구조 개선
