+++ components/Chart/RadialChart/subtitle.tsx
@@ -1,0 +1,21 @@
+import { cn } from "@/utils/cn";
+
+export function RadialChartSubtitle({
+  color,
+  className,
+  children,
+  ...props
+}: { color: string } & React.HTMLAttributes<HTMLSpanElement>) {
+  return (
+    <span
+      className={cn(
+        "text-sm font-medium",
+        className
+      )}
+      style={{ color }}
+      {...props}
+    >
+      {children}
+    </span>
+  )
+}
